# Conflicts:
#	src/net/sf/odinms/server/MapleItemInformationProvider.java

+++ src/net/sf/odinms/net/channel/handler/ChangeMapSpecialHandler.java
@@ -9,11 +9,10 @@ import net.sf.odinms.tools.MaplePacketCreator;
 import net.sf.odinms.tools.data.input.SeekableLittleEndianAccessor;
 
 public class ChangeMapSpecialHandler extends AbstractMaplePacketHandler {
-
     @Override
     public void handlePacket(SeekableLittleEndianAccessor slea, MapleClient c) {
         try {
-            c.getPlayer().setLastKillOnMap((long) 0);
+            c.getPlayer().setLastKillOnMap(0L);
             c.getPlayer().setLastDamageSource(null);
             c.getPlayer().resetAfkTime();
             slea.readByte();
